+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Tips!
+
 ## [0.2.0] 2023-11-27
 
 ### Added